tutorial/requirements.txt
Django==2.1.4
djangorestframework==3.11.0
Pygments==2.6.1
pylint==2.5.2
